fix: Detect keycode for '?' correctly.
Previously, the '?' was covered under the default case, but
then the help screen appeared when it should not, for example,
when opening the developer tools.

Upon further inspection, it became clear that the wrong key code
was used (63 instead of 191), and after using 191, it showed that
player.handleInput was not guarded against being passed an "undefined"
keycode. (when pressing Shift+/ to reach '?', the help screen appeared, but
immediately disappeared again after releasing the Shift-key).