refactor: pause functionality now in the game loop
Not sure whether the extensive refactoring in engine.js
was necessary. The crux is that the engine needs to be
accessible in player (engine.togglePaused()), and that was not
the case.

Currently, too much of the Engine is publicly accessible (though not in the
global namespace) for my taste.

Perhaps it would have been possible to refrain from most of the refactoring
had I known beforehand about "bind". I found out late about it and it
was the key to get some callbacks to work.